flask/requirements.txt
flask==0.12.2
flask-sqlalchemy==2.3.1
gunicorn==19.7.1
psycopg2==2.7.3.1
